requirements.txt
litellm
deepgram-sdk
playsound3
langchain-groq 
langchain_google_genai==1.0.1
langchain_chroma
chromadb
Pillow
pyperclip
pyaudio
python-dotenv
instructor 
pydantic
colorama
tavily-python
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client